refactor handler into backend and stream
